feat: topolog h-cube dim env var: SP1_TOPOLOGY_DIM

+++ crates/core/executor/src/events/precompiles/topology.rs
@@ -5,7 +5,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Clone, Serialize, Deserialize, DeepSizeOf, Default)]
 pub struct TopologicalRouteEvent {
     /// The current node ID in the graph.
-    pub current_node: u32,
+    pub current_node: u64,
     /// The next valid node ID in the graph.
-    pub next_node: u32,
+    pub next_node: u64,
 }
